Add shared implementation of dispatching logic
