+++ src/Microsoft.EntityFrameworkCore.Tools.Cli/project.json
@@ -17,6 +17,7 @@
     "Microsoft.DotNet.Cli.Utils": "1.0.0-*",
     "Microsoft.DotNet.ProjectModel": "1.0.0-*",
     "Microsoft.Extensions.CommandLineUtils": "1.0.0-*",
+    "Microsoft.Extensions.DotnetToolDispatcher.Sources": "1.0.0-*",
     "Microsoft.EntityFrameworkCore.Tools.Core": "1.0.0-*"
   },
   "frameworks": {
